New features incomplete
I'm working on new features to set scheduling of the events returned by the EMU-2 which aren't working at the moment.  There's some added output during startup, but after that it will still report JSON data as events are received.

Also added in some command line support for a -p parameter to specify the serial port in use instead of having it hard-coded in the code.